+++ packages/core/src/types/get-all-object-paths.ts
@@ -25,6 +25,8 @@ export type GetAllObjectPathsEntries<
       path: string;
       type: any;
     }
+  : Record<string | number, any> extends O
+  ? any
   :
       | (P extends '' ? never : ObjectTypeEntry<P, O>)
       | (O extends unknown[]